DataWearhouse/requirements.txt
mysql-connector-python
schedule
requests